fix: circular imports and more shenanigans
