+++ util/constants.py
@@ -1,0 +1,1 @@
+version = "3.2.1"
